Extract models to extend UserPost
