+++ app/src/main/java/example/com/cmsandroidsimulation/models/Complaint.java
@@ -2,20 +2,9 @@ package example.com.cmsandroidsimulation.models;
 
 import java.util.Date;
 
-public class Complaint{
-    private final String username;
-    private final String content;
+public class Complaint extends UserPost{
     public Complaint(String username, String content)
     {
-        this.username = username;
-        this.content = content;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public String getUsername() {
-        return username;
+        super(username, null, content);
     }
 }
