+++ app/src/main/java/example/com/cmsandroidsimulation/models/EventInfo.java
@@ -3,14 +3,11 @@ package example.com.cmsandroidsimulation.models;
 import java.util.ArrayList;
 import java.util.Date;
 
-public class EventInfo {
+public class EventInfo extends UserPost{
 
     // Might change to builder pattern later.
 
     private final int id;
-    private final String author;
-    private final String title;
-    private final String details;
     private final float rating;
     private final ArrayList<EventComment> comments;
     private final Date eventStartDateTime;
@@ -20,10 +17,8 @@ public class EventInfo {
                      Date eventStartDateTime, Date eventEndDateTime,
                      ArrayList<EventComment> comments)
     {
+        super(author, title, details);
         this.id = id;
-        this.author = author;
-        this.title = title;
-        this.details = details;
         this.rating = rating;
         this.comments = comments;
         this.eventStartDateTime = eventStartDateTime;
@@ -38,14 +33,6 @@ public class EventInfo {
         return rating;
     }
 
-    public String getDetails() {
-        return details;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
     public ArrayList<EventComment> getComments() {
         return comments;
     }
@@ -56,10 +43,5 @@ public class EventInfo {
 
     public Date getEventStartDateTime() {
         return eventStartDateTime;
-    }
-
-
-    public String getAuthor() {
-        return author;
     }
 }
